requirements.txt
fastapi
uvicorn[standard]
paho-mqtt
numpy
PyYAML
pymodbus
aiofiles
